Small clean-up.
